Changing a little bit of transitioning logic
I've changed the logic such that only each viewport owns a subset of the
school of fish. So, when a fish that they own transitions to another
viewport they notify everyone. The benefit of this is so that there's
never a possibility of two viewports claiming ownership of a fish. This
makes the transitions not look as clean, but we've eliminated the
jerkiness from the original version